update social icons color
